requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
easyocr>=1.7.0
firebase-admin>=6.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
torch>=2.0.0
